+++ source/common/utils/environmentCheckers.js
@@ -5,6 +5,7 @@ import {
   LINUX,
   MAC_OS,
   MAINNET,
+  MAINNET_EAG,
   MAINNET_FLIGHT,
   PRODUCTION,
   SHELLEY_TESTNET,
@@ -30,7 +31,7 @@ export const evaluateNetwork = (network: ?string) => {
   if (network === QA || network === NIGHTLY || network === ITN_SELFNODE) {
     currentNetwork = ITN_REWARDS_V1;
   }
-  if (network === MAINNET_FLIGHT) {
+  if (network === MAINNET_EAG || network === MAINNET_FLIGHT) {
     currentNetwork = MAINNET;
   }
   if (network === SHELLEY_TESTNET_V6) {
@@ -44,10 +45,12 @@ export const getBuildLabel = (
   network: string,
   currentNodeEnv: string,
   isFlight: boolean,
+  isMainnetEAG: boolean,
   version: string
 ) => {
   const networkLabel = isFlight ? 'Flight' : networkPrettyNames[network];
-  let buildLabel = `Daedalus ${networkLabel} (${version}#${build})`;
+  const networkLabelSuffix = isMainnetEAG ? ' - EAG' : '';
+  let buildLabel = `Daedalus ${networkLabel}${networkLabelSuffix} (${version}#${build})`;
   if (!checkIsProduction(currentNodeEnv))
     buildLabel += ` ${upperFirst(currentNodeEnv)}`;
   return buildLabel;