+++ source/main/environment.js
@@ -7,6 +7,7 @@ import {
   DEVELOPMENT,
   OS_NAMES,
   MAINNET,
+  MAINNET_EAG,
   MAINNET_FLIGHT,
   SHELLEY_TESTNET,
   SHELLEY_TESTNET_V6,
@@ -37,7 +38,9 @@ import {
 // environment variables
 const CURRENT_NODE_ENV = process.env.NODE_ENV || DEVELOPMENT;
 let RAW_NETWORK =
-  process.env.NETWORK === MAINNET_FLIGHT ? MAINNET : process.env.NETWORK || '';
+  process.env.NETWORK === MAINNET_EAG || process.env.NETWORK === MAINNET_FLIGHT
+    ? MAINNET
+    : process.env.NETWORK || '';
 if (process.env.NETWORK === SHELLEY_TESTNET_V6) {
   RAW_NETWORK = SHELLEY_TESTNET;
 }