+++ source/renderer/app/containers/TopBarContainer.js
@@ -39,9 +39,10 @@ export default class TopBarContainer extends Component<Props> {
       ? menuIconOpened
       : menuIconClosed;
     const leftIcon = showSubMenuToggle ? leftIconSVG : null;
-    const testnetLabel = !isMainnet ? (
-      <WalletTestEnvironmentLabel network={network} />
-    ) : null;
+    const testnetLabel =
+      global.isMainnetEAG || !isMainnet ? (
+        <WalletTestEnvironmentLabel network={network} />
+      ) : null;
 
     const onWalletAdd = () => {
       actions.router.goToRoute.trigger({